Vueified [wip broken]
