+++ package.json
@@ -38,8 +38,8 @@
     "font-awesome-filetypes": "^1.0.1"
   },
   "peerDependencies": {
-    "vue": "^1.0.24",
-    "vuex": "2.0.0-rc.1"
+    "vue": "^2.0.0-rc.7",
+    "vuex": "2.0.0-rc.6"
   },
   "devDependencies": {
     "autoprefixer": "^6.3.6",
@@ -59,8 +59,8 @@
     "postcss-easy-import": "^1.0.1",
     "postcss-nested": "^1.0.0",
     "postcss-simple-vars": "^2.0.0",
-    "vue": "^1.0.25",
-    "vuex": "2.0.0-rc.1",
+    "vue": "^2.0.0-rc.7",
+    "vuex": "2.0.0-rc.6",
     "watchify": "^3.7.0"
   },
   "babel": {
