+++ src/components/media/media-table.vue
@@ -1,28 +1,29 @@
+<template>
+    <table
+        v-el:container
+        v-sortable
+        handle=".js-handle"
+        class="media__table"
+    >
+        <tbody v-for="media in orderedMedia">
+            <tr
+                is="media-row"
+                :media="media"
+                :data="data"
+                :options="options"
+                class="media__row"
+                :class="{ '-is-disabled': media.markedForRemoval }"
+            ></tr>
+        </tbody>
+    </table>
+</template>
+
+<script>
 import MediaRow from './media-row';
 import Sortable from '../sortable/sortable';
 import { sortBy } from 'lodash';
 
 export default {
-
-    template: `
-        <table
-            v-el:container
-            v-sortable
-            handle=".js-handle"
-            class="media__table"
-        >
-            <tbody v-for="media in orderedMedia">
-                <tr
-                    is="media-row"
-                    :media="media"
-                    :data="data"
-                    :options="options"
-                    class="media__row"
-                    :class="{ '-is-disabled': media.markedForRemoval }"
-                ></tr>
-            </tbody>
-        </table>
-    `,
 
     props: ['collection', 'media', 'options', 'data'],
 
@@ -56,3 +57,4 @@ export default {
     },
 
 };
+</script>
