+++ src/components/media/thumb.vue
@@ -1,28 +1,29 @@
+<template>
+    <a
+        :href="originalUrl"
+        target="blender_media"
+        tabIndex="-1"
+        class="media__thumb"
+    >
+        <img
+            v-if="isImage"
+            :src="thumbUrl"
+            class="media__thumb__image"
+          />
+        <span
+            v-else
+            class="media__thumb__file"
+        >
+            <i :class="['fa', icon, 'media__thumb__file__icon']"></i>
+        </span>
+    </a>
+</template>
+
+<script>
 import getClassNameForExtension from 'font-awesome-filetypes';
 import { includes } from 'lodash';
 
 export default {
-
-    template: `
-        <a
-            :href="originalUrl"
-            target="blender_media"
-            tabIndex="-1"
-            class="media__thumb"
-        >
-            <img
-                v-if="isImage"
-                :src="thumbUrl"
-                class="media__thumb__image"
-              />
-            <span
-                v-else
-                class="media__thumb__file"
-            >
-                <i :class="['fa', icon, 'media__thumb__file__icon']"></i>
-            </span>
-        </a>
-    `,
 
     props: ['media'],
 
@@ -45,3 +46,4 @@ export default {
     },
 
 };
+</script>
