+++ src/modules/errors.js
@@ -6,15 +6,12 @@ const state = {
 };
 
 export const mutations = {
-
     addError(state, { collection, message }) {
         Vue.set(state.errors, collection, message);
     },
-
     clearErrors(state, { collection }) {
         Vue.delete(state.errors, collection);
     },
-
 };
 
 export const actions = {
