+++ src/modules/media.js
@@ -7,30 +7,26 @@ const state = {
 };
 
 export const mutations = {
-
     addMedia(state, { media }) {
-
-        if (!Array.isArray(media)) {
+        if (! Array.isArray(media)) {
             media = [media];
         }
 
         media.forEach(media => Vue.set(state.media, media.id, media));
     },
-
     renameMedia(state, { id, name }) {
-
-        if(!state.media[id]) return;
-
+        if(!state.media[id]) {
+            return;
+        }
         state.media[id].name = name;
     },
-
     markMediaForRemoval(state, { id }) {
-
-        if(!state.media[id]) return;
+        if(! state.media[id]) {
+            return;
+        }
 
         Vue.set(state.media[id], 'markedForRemoval', true);
     },
-
     markCollectionForRemoval(state, { collection }) {
         forIn(state.media, (media) => {
             if (media.collection === collection) {
@@ -38,19 +34,17 @@ export const mutations = {
             }
         });
     },
-
     restoreMedia(state, { id }) {
-
-        if(!state.media[id]) return;
+        if(! state.media[id]) {
+            return;
+        }
 
         Vue.set(state.media[id], 'markedForRemoval', false);
     },
-
     replaceMedia(state, { collection, media }) {
         mutations.clearCollection(state, { collection });
         mutations.addMedia(state, { media });
     },
-
     clearCollection(state, { collection }) {
         forIn(state.media, (media, id) => {
             if (media.collection === collection) {
@@ -58,16 +52,15 @@ export const mutations = {
             }
         });
     },
-
     setMediaOrder(state, { order }) {
         forIn(order, (order, mediaId) => {
             state.media[mediaId].orderColumn = order;
         });
     },
-
     updateCustomProperty(state, { id, property, value }) {
-
-        if(!state.media[id]) return;
+        if(! state.media[id]) {
+            return;
+        }
 
         const [ namespace, key ] = property.split('.');
 
@@ -82,7 +75,6 @@ export const mutations = {
 
         Vue.set(state.media[id].customProperties[namespace], key, value);
     },
-
 };
 
 export const actions = {
